+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `frame` parameter to `hopkins()` and `hopkins_test()` for specifying
+  a custom rectangular sampling frame by its bounds.
+
 ## [0.4.0] - 2026-02-16
 
 ### Fixed